Completed MobileSettings component, moved generic components to utilities folder
