+++ frontend/src/components/utilities/SwitchTheme.jsx
@@ -1,9 +1,9 @@
 import PropTypes from "prop-types";
 
-import SwitchLightIcon from "./navsettings/SwitchLightIcon";
-import SwitchDarkIcon from "./navsettings/SwitchDarkIcon";
+import SwitchLightIcon from "../navsettings/SwitchLightIcon";
+import SwitchDarkIcon from "../navsettings/SwitchDarkIcon";
 
-import styles from "../css/SwitchTheme.module.css";
+import styles from "../../css/SwitchTheme.module.css";
 
 function SwitchTheme({ isToggled, onToggle, showIcons = true }) {
   return (
